requirements.txt
chromadb~=0.5.13
streamlit~=1.40.0

numpy~=1.26.4
typing_extensions~=4.12.2
pillow~=10.4.0
pip~=24.3.1
attrs~=24.2.0
wheel~=0.41.2
tornado~=6.4.1
Jinja2~=3.1.4
filelock~=3.16.1
requests~=2.32.3
rich~=13.8.1
colorama~=0.4.6
PyYAML~=6.0.2
scipy~=1.14.1
packaging~=24.1
pytz~=2024.2
overrides~=7.7.0
regex~=2024.9.11
threadpoolctl~=3.5.0
matplotlib~=3.8.3
mpmath~=1.3.0
sympy~=1.13.3
decorator~=5.1.1
tqdm~=4.66.5
urllib3~=2.2.3
fsspec~=2024.6.1
yarl~=1.13.1
MarkupSafe~=2.1.5
joblib~=1.4.2
certifi~=2024.8.30
scikit-learn~=1.5.2
pandas==1.5.3
networkx~=3.3
idna~=3.10
pyparsing~=3.1.4

Pygments~=2.18.0
nltk~=3.9.1
fastapi~=0.115.0
pydantic~=2.11.4
starlette~=0.40.0
uvicorn~=0.31.1
bitsandbytes~=0.44.0
langchain~=0.3.3
distro~=1.9.0
python-dotenv~=1.0.1
pandas==1.5.3
bs4~=0.0.2

neurokit2~=0.2.10
langchain-ollama==0.2.0

cffi~=1.17.1
defusedxml~=0.7.1
ipython~=8.29.0
jupyter~=1.1.1
ipywidgets~=8.1.5
notebook~=7.2.2
torch~=2.4.1
setuptools~=68.2.0
platformdirs~=4.3.6
onnxruntime~=1.19.2
torchvision~=0.19.1
psutil~=6.0.0
murmurhash~=1.0.10
datasets~=3.0.1
zipp~=3.20.2
build~=1.2.2.post1
tokenizers~=0.20.1
safetensors~=0.4.5
transformers~=4.46.3
accelerate~=0.34.2
sentencepiece~=0.2.0
beautifulsoup4~=4.12.3
pytesseract~=0.3.13
opencv-python~=4.10.0.84
aiohttp~=3.10.7
bleak~=0.22.3
pytest~=8.3.3
pyshimmer~=0.4.0
pyserial~=3.5
toml~=0.10.2
typing~=3.7.4.3
mistralai~=0.4.2